web-spring-shop/src/main/java/com/epam/learn/shchehlov/webspringshop/controller/UserController.java
package com.epam.learn.shchehlov.webspringshop.controller;

import com.epam.learn.shchehlov.webspringshop.dto.UserDto;
import com.epam.learn.shchehlov.webspringshop.entity.User;
import com.epam.learn.shchehlov.webspringshop.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.BindingResult;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;

@RestController
public class UserController {

    private final UserService userService;

    @Autowired
    public UserController(UserService userService) {
        this.userService = userService;
    }

    @PostMapping(value = "/registration")
    public ResponseEntity<?> createNewUser(@Valid UserDto userDto, BindingResult bindingResult) {
        User userExists = userService.findUserByLogin(userDto.getLogin());

        if (userExists != null) {
            bindingResult.rejectValue("login", "error.user",
                            "There is already a user registered with the login provided");
        }

        if (bindingResult.hasErrors()) {
            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
        } else {
            return new ResponseEntity<>(userService.createUser(userDto), HttpStatus.CREATED);
        }
    }
}
